backend/requirements.txt
flask
flask-cors
numpy
joblib
scikit-learn
mysql-connector-python
